Implement test for playing cards
For now we do nothing with them, that will be for round resolve phase.

Usage of R3 will be done once we reach monobehaviours, as those will need them to react to the data.

For now continue implementing all the data part first, then implement the rest of the game around it.