Resume arranging UI for player and Table
Cards and suits are updated now on PlayerView, general game now also updates roundIndex.

Use ReactiveProperties from R3 and IObserver / observable depending on use case.

All of this needs to be started with finishing the validation test for the round itself to work (call data as usual and fetch view to see if it holds proper data?)